requirements.txt
fastapi
uvicorn
pydantic
psycopg2-binary
reportlab
qrcode[pil]
jinja2
python-multipart